Deploys first working prototype of Mortybot
Morty understands 'help' and 'roll' commands. Code tree reorganized to move
legacy commands to a different part of the tree (and thus not evaluated) as
part of the set of available commands.

Debug statements removed that were originally inserted to troubleshoot Heroku
environment variable issues associated with slash token authentication.